# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TwinBlueprint Server API",
    "description": "REST API for TwinBlueprint — leads, bids, projects, campaigns, analytics, and admin auth.",
    "version": "3.0.0",
    "contact": {
      "name": "TwinBlueprint",
      "email": "twinblueprints@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token. Can also be sent via cookie named 'token'."
      }
    },
    "schemas": {
...
      "Bid": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "project": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "phase": {
            "type": "string",
            "enum": [
              "RFP Review",
              "Technical Eval",
              "Shortlist"
            ]
          },
          "deadline": {
            "type": "string",
            "description": "Date string (YYYY-MM-DD)"
          },
          "suppliers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Supplier names; empty when unconfirmed."
          },
          "value": {
            "type": "number",
            "nullable": true,
            "description": "Actual bid amount. Zero is valid; null means unconfirmed. Independent of lead.project_size."
          },
          "lead_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Existing lead ID. At most one bid per linked lead."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "pattern": "^[A-Z]{3}$",
            "description": "Explicit source ISO currency; null when unknown. Never inferred from geography."
          }
        },
        "required": [
          "id",
          "project",
          "client",
          "phase",
          "deadline",
          "suppliers",
          "created_at"
        ]
      },
      "CreateBidInput": {
        "type": "object",
        "properties": {
          "project": {
            "type": "string",
            "example": "Hudson Yards Phase 4"
          },
          "client": {
            "type": "string",
            "example": "Related Companies"
          },
          "phase": {
            "type": "string",
            "enum": [
              "RFP Review",
              "Technical Eval",
              "Shortlist"
            ]
          },
          "deadline": {
            "type": "string",
            "format": "date",
            "description": "A real calendar date in YYYY-MM-DD format. Slash dates and impossible dates are rejected."
          },
          "suppliers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Supplier names; empty when unconfirmed."
          },
          "value": {
            "type": "number",
            "nullable": true,
            "description": "Actual bid amount. Zero is valid; null means unconfirmed. Independent of lead.project_size."
          },
          "lead_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Existing lead ID. At most one bid per linked lead."
          },
          "status": {
            "type": "string",
            "default": "Active",
            "description": "Any nonempty string is currently accepted. Defaults to Active on creation; active_bids includes only status Active (case-insensitive). No fixed bid-status enum is enforced."
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "pattern": "^[A-Z]{3}$",
            "description": "Explicit source ISO currency; null when unknown. Never inferred from geography."
          }
        },
        "required": [
          "project",
          "client",
          "phase",
          "deadline"
        ]
      },
      "UpdateBidInput": {
        "type": "object",
        "properties": {
          "project": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "phase": {
            "type": "string",
            "enum": [
              "RFP Review",
              "Technical Eval",
              "Shortlist"
            ]
          },
          "deadline": {
            "type": "string",
            "format": "date",
            "description": "A real calendar date in YYYY-MM-DD format. Slash dates and impossible dates are rejected."
          },
          "suppliers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Supplier names; empty when unconfirmed."
          },
          "value": {
            "type": "number",
            "nullable": true,
            "description": "Actual bid amount. Zero is valid; null means unconfirmed. Independent of lead.project_size."
          },
          "lead_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Existing lead ID. At most one bid per linked lead."
          },
          "status": {
            "type": "string",
            "description": "Any nonempty string is currently accepted. Defaults to Active on creation; active_bids includes only status Active (case-insensitive). No fixed bid-status enum is enforced."
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "pattern": "^[A-Z]{3}$",
            "description": "Explicit source ISO currency; null when unknown. Never inferred from geography."
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "string",
            "description": "Only included in development"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "region": "Region is required"
            }
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/bids": {
      "post": {
        "tags": [
          "Bids"
        ],
        "summary": "Create a bid",
        "description": "Admin only. Create a bid using project, client, phase, and deadline. Supply lead_id to link an existing lead without changing its status or project_size. value defaults to null and suppliers to []. Duplicate lead links return 409; update the existing bid instead.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBidInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bid created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bid": {
                          "$ref": "#/components/schemas/Bid"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bid or linked lead not found."
          },
          "409": {
            "description": "The lead already has a bid. Update that bid instead."
          }
        }
      },
      "get": {
        "tags": [
          "Bids"
        ],
        "summary": "List all bids",
        "description": "Supports pagination.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bids": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Bid"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/bids/{id}": {
      "get": {
        "tags": [
          "Bids"
        ],
        "summary": "Get bid by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bid": {
                          "$ref": "#/components/schemas/Bid"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bid not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Bids"
        ],
        "summary": "Update a bid",
        "description": "Admin only. Update one or more bid fields, including value (number or null), suppliers (array of names), and lead_id. Omitted fields are preserved. Does not change the linked lead's status or project_size.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBidInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bid updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "bid": {
                          "$ref": "#/components/schemas/Bid"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bid or linked lead not found."
          },
          "409": {
            "description": "The lead already has a bid. Update that bid instead."
          }
        }
      },
      "delete": {
        "tags": [
          "Bids"
        ],
        "summary": "Delete a bid",
        "description": "Admin only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bid deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bids/{id}/suppliers/{supplierId}": {
      "put": {
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Link existing supplier to bid (admin, idempotent)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link saved"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Bid or supplier not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Suppliers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Unlink supplier from bid (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link saved"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Bid or supplier not found"
          }
        }
      }
    },
...
  }
}